% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_wto_crawler}
\alias{new_wto_crawler}
\title{new_wto_crawler}
\usage{
new_wto_crawler(.verbose = FALSE, .proxy = FALSE)
}
\arguments{
\item{.verbose}{TRUE or FALSE}
}
\value{
list
}
\description{
new_wto_crawler
}
