% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_post.R
\name{request_wto_post.wto}
\alias{request_wto_post.wto}
\title{request_wto_post}
\usage{
\method{request_wto_post}{wto}(x, params, verbose = FALSE,
  proxy = FALSE, timeout = 60)
}
\arguments{
\item{x}{.}

\item{params}{.}

\item{verbose}{.}

\item{proxy}{.}

\item{timeout}{.}
}
\value{
.
}
\description{
request_wto_post
}
