% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tariff_download.R
\name{tariff_downloader}
\alias{tariff_downloader}
\title{tariff_downloader}
\usage{
tariff_downloader(import, export, hs2, hs4)
}
\arguments{
\item{import}{.}

\item{export}{.}

\item{hs2}{.}

\item{hs4}{.}
}
\value{
.
}
\description{
tariff_downloader
}
